Split the field files
