+++ src/wtforms/fields/__init__.py
@@ -1,7 +1,11 @@
-from wtforms.fields.core import *
+from wtforms.fields.choices import *
+from wtforms.fields.choices import SelectFieldBase
 from wtforms.fields.core import Field
 from wtforms.fields.core import Flags
 from wtforms.fields.core import Label
-from wtforms.fields.core import SelectFieldBase
+from wtforms.fields.datetime import *
+from wtforms.fields.form import *
+from wtforms.fields.list import *
+from wtforms.fields.numeric import *
 from wtforms.fields.simple import *
 from wtforms.utils import unset_value as _unset_value
